+++ auction-house-v2/programs/auction-house-v2/src/state.rs
@@ -21,7 +21,7 @@ use anchor_lang::prelude::*;
 pub struct AuctionHouseV2Data {
     pub authority: Pubkey,
     pub treasury_mint: Pubkey,
-    pub seller_fee_basis_points: u64,
+    pub seller_fee_basis_points: u16,
     pub requires_sign_off: bool,
     pub treasury_account: Pubkey,
     pub treasury_withdrawal_account: Pubkey,
